pyproject.toml
[tool.poetry]
name = "gitlab_kanban"
version = "0.1.0"
description = ""
authors = ["ryonak <nakagamiryo0901@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
polars = "^0.17.12"
DateTime = "^5.1"
plotly = "^5.14.1"
pytest = "^7.3.1"
requests = "^2.30.0"
python-gitlab = "^3.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
